Added heartbeat. Modified tests to work okay with heartbeat

+++ testing/functions.t
@@ -1,8 +1,11 @@
 #!/bin/sh
 
 setup(){
-  export TF_HOME=/tmp/tf.$$
-  (cd ../;make install prefix=$TF_HOME)
+  [ -z $TF_HOME ] && export TF_HOME=/tmp/tf.$$
+  [ -d $TF_HOME ] || mkdir $TF_HOME
+  if [ ! -e $TF_HOME/bin/tfrun ] ; then
+    (cd ../;make install prefix=$TF_HOME)
+  fi
   export TFILE=test.faa
   ME=`pwd`/$0
   cp $TFILE $TF_HOME
@@ -10,7 +13,7 @@ setup(){
 }
 
 cleanup(){
-  for f in data.dump  error.test.faa  fastrecovery.test.faa  log.test.faa  progress.test.faa  status.new  test.err  test.out tf.fastrecovery progress.$TFILE test.args tf.err tf.log tf.pid
+  for f in data.dump  error.test.faa  fastrecovery.test.faa  log.test.faa  progress.test.faa done.test.faa status.new  test.err  test.out tf.fastrecovery progress.$TFILE test.args tf.err tf.log tf.pid test2.out test2.err
   do
     [ -e $f ] && rm $f
   done
